chore(asm): move ASM Django snapshot tests to specific appsec test file (#5145)
1. Move two snapshots tests from
`tests/contrib/django/test_django_snapshots` to
`tests/contrib/django/test_django_appsec_snapshots`

- test_appsec_enabled
- test_appsec_enabled_attack
- test_request_ipblock_nomatch_200
- test_request_ipblock_match_403
- test_request_ipblock_match_403_json

2. Rename 5 snapshot files to take the previous change into account.

**Motivation**: many changes ahead in the ASM features will require to
update those snapshots and add more tests.

## Checklist

- [x] Change(s) are motivated and described in the PR description.
- [x] Testing strategy is described if automated tests are not included
in the PR.
- [x] Risk is outlined (performance impact, potential for breakage,
maintainability, etc).
- [x] Change is maintainable (easy to change, telemetry, documentation).
- [x] [Library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/contributing.html#Release-Note-Guidelines)
are followed.
- [x] Documentation is included (in-code, generated user docs, [public
corp docs](https://github.com/DataDog/documentation/)).
- [x] Author is aware of the performance implications of this PR as
reported in the benchmarks PR comment.

## Reviewer Checklist

- [x] Title is accurate.
- [x] No unnecessary changes are introduced.
- [x] Description motivates each change.
- [x] Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes unless absolutely necessary.
- [x] Testing strategy adequately addresses listed risk(s).
- [x] Change is maintainable (easy to change, telemetry, documentation).
- [x] Release note makes sense to a user of the library.
- [x] Reviewer is aware of, and discussed the performance implications
of this PR as reported in the benchmarks PR comment.
